feat: Implement GetWeather
Implement GetWeather from the presentation to the
data layer.

Presentation: Add a BlocProvider to main.dart and
use a button on the WeatherPage to trigger a fetch.

Domain: Complete the getWeather usecase with
latitude and logitude.

Data: Complete the getWeather method on the
repository with latitude, longitude and error
handling.
Link the remote data source get method with the
api client to fetch weather.

Add tests for the usecase, repository and
data source.