feat(city): Implement local data source
Implement a local data source to store registered
cities.
Create getSavedCities to fetch a list of saved
cities.
Create saveCity to save a new city.
Create removeCity to remove a city from the list.
Add a toJson method to CityModel.

Create a String hive box and add it to dependency
injection.

Add unit tests for the local data source.