refactor(city): Change GetSavedCities as GetSavedCitiesWeather
Refactor GetSavedCities usecase to not only return
a city list, but a city and its weather forecast.
Create a CityWeather entity which is a combination
of a city and a weather entities.

Rename SavedCitiesBloc variable cities as
cityWeatherList.

Display the current temperature on WeatherPage
list items.

Rename the usecase as GetSavedCitiesWeather.
Refactor unit test.